backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
PyMySQL==1.1.1
cryptography==42.0.8
python-dotenv==1.0.1
requests==2.32.3
